python/requirements.txt
pyserial
requests